+++ Unresize/unresize_impl.h
@@ -1,14 +1,15 @@
 #ifndef ZIMG_UNRESIZE_IMPL_H_
 #define ZING_UNRESIZE_IMPL_H_
 
+#include <cstddef>
 #include "Common/osdep.h"
 #include "bilinear.h"
 
 namespace zimg {;
 namespace unresize {;
 
-FORCE_INLINE inline void filter_scanline_h_forward(const BilinearContext &ctx, const float * RESTRICT src, float * RESTRICT tmp, int src_stride,
-                                                   int i, int j_begin, int j_end)
+FORCE_INLINE inline void filter_scanline_h_forward(const BilinearContext &ctx, const float * RESTRICT src, float * RESTRICT tmp,
+                                                    ptrdiff_t src_stride, ptrdiff_t i, ptrdiff_t j_begin, ptrdiff_t j_end)
 {
 	const float *c = ctx.lu_c.data();
 	const float *l = ctx.lu_l.data();
@@ -16,12 +17,12 @@ FORCE_INLINE inline void filter_scanline_h_forward(const BilinearContext &ctx, c
 	float z = j_begin ? src[i * src_stride + j_begin - 1] : 0;
 
 	// Matrix-vector product, and forward substitution loop.
-	for (int j = j_begin; j < j_end; ++j) {
+	for (ptrdiff_t j = j_begin; j < j_end; ++j) {
 		const float *row = ctx.matrix_coefficients.data() + j * ctx.matrix_row_stride;
-		int left = ctx.matrix_row_offsets[j];
+		ptrdiff_t left = ctx.matrix_row_offsets[j];
 
 		float accum = 0;
-		for (int k = 0; k < ctx.matrix_row_size; ++k) {
+		for (ptrdiff_t k = 0; k < ctx.matrix_row_size; ++k) {
 			accum += row[k] * src[i * src_stride + left + k];
 		}
 
@@ -30,33 +31,33 @@ FORCE_INLINE inline void filter_scanline_h_forward(const BilinearContext &ctx, c
 	}
 }
 
-FORCE_INLINE inline void filter_scanline_h_back(const BilinearContext &ctx, const float * RESTRICT tmp, float * RESTRICT dst, int dst_stride,
-                                                int i, int j_begin, int j_end)
+FORCE_INLINE inline void filter_scanline_h_back(const BilinearContext &ctx, const float * RESTRICT tmp, float * RESTRICT dst, 
+                                                ptrdiff_t dst_stride,ptrdiff_t i, ptrdiff_t j_begin, ptrdiff_t j_end)
 {
 	const float *u = ctx.lu_u.data();
 	float w = j_begin < ctx.dst_width ? dst[i * dst_stride + j_begin] : 0;
 
 	// Backward substitution.
-	for (int j = j_begin; j > j_end; --j) {
+	for (ptrdiff_t j = j_begin; j > j_end; --j) {
 		w = tmp[j - 1] - u[j - 1] * w;
 		dst[i * dst_stride + j - 1] = w;
 	}
 }
 
-FORCE_INLINE inline void filter_scanline_v_forward(const BilinearContext &ctx, const float * RESTRICT src, float * RESTRICT dst, int src_stride, int dst_stride,
-                                                   int i, int j_begin, int j_end)
+FORCE_INLINE inline void filter_scanline_v_forward(const BilinearContext &ctx, const float * RESTRICT src, float * RESTRICT dst,
+                                                   ptrdiff_t src_stride, ptrdiff_t dst_stride, ptrdiff_t i, ptrdiff_t j_begin, ptrdiff_t j_end)
 {
 	const float *c = ctx.lu_c.data();
 	const float *l = ctx.lu_l.data();
 
 	const float *row = ctx.matrix_coefficients.data() + i * ctx.matrix_row_stride;
-	int top = ctx.matrix_row_offsets[i];
+	ptrdiff_t top = ctx.matrix_row_offsets[i];
 
-	for (int j = j_begin; j < j_end; ++j) {
+	for (ptrdiff_t j = j_begin; j < j_end; ++j) {
 		float z = i ? dst[(i - 1) * dst_stride + j] : 0;
 
 		float accum = 0;
-		for (int k = 0; k < ctx.matrix_row_size; ++k) {
+		for (ptrdiff_t k = 0; k < ctx.matrix_row_size; ++k) {
 			accum += row[k] * src[(top + k) * src_stride + j];
 		}
 
@@ -65,11 +66,11 @@ FORCE_INLINE inline void filter_scanline_v_forward(const BilinearContext &ctx, c
 	}
 }
 
-FORCE_INLINE inline void filter_scanline_v_back(const BilinearContext &ctx, float * RESTRICT dst, int dst_stride, int i, int j_begin, int j_end)
+FORCE_INLINE inline void filter_scanline_v_back(const BilinearContext &ctx, float * RESTRICT dst, ptrdiff_t dst_stride, ptrdiff_t i, ptrdiff_t j_begin, ptrdiff_t j_end)
 {
 	const float *u = ctx.lu_u.data();
 
-	for (int j = j_begin; j < j_end; ++j) {
+	for (ptrdiff_t j = j_begin; j < j_end; ++j) {
 		float w = i < ctx.dst_width ? dst[i * dst_stride + j] : 0;
 
 		w = dst[(i - 1) * dst_stride + j] - u[i - 1] * w;
